refactor: Implement useSingUp hook
